Add Heartrate and Stress to API database
The commit introduces HeartRate and Stress models to the API database. It includes the creation of associated constants, structs and methods for these models in query.sql.go file and the addition of HeartRate and Stress types in models.go. Additionally, it involves adjustments to the trigger functions of both spo2.sql and stress.sql and new queries are added to the query.sql file.