v0.0.5
Refactor data types for int fields in DB models

The data types for all integer fields in the database queries, models, and configurations have been changed from int32 to int to match with the database schema. This update provides consistency between the application and the database which will slim the chance for errors.