v0.0.8
Update GetSleep queries to return :many even though they would ever only return at most one row.

I just simply don't like that if you queried for an ID that wasn't in the database it results in an error and you'd need to look at the text of the error to determine that it simply was a record not found.

Now we would get an empty array.