Update loop count and data type for stress duration
This commit makes two important changes: Firstly, it replaces the hard-coded count of 50 with the numberOfRecords variable in the data generation loops of the heartrate, spo2, and stress examples. This makes the number of iterations flexible and centrally configurable. Secondly, it changes the data type of the HighStressDuration field in the Stress struct and in relevant database structures from int to int64, allowing for more precise tracking of stress duration.